ai-services/requirements.txt
# AI服务依赖包

# 核心框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# 机器学习
tensorflow==2.13.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# 文档处理
pytesseract==0.3.10
Pillow==10.1.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.11.7
PyMuPDF==1.26.4
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.1.3
xlrd==2.0.1
xlwt==1.3.0
python-magic==0.4.27
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据库连接
pymongo==4.6.0
redis==5.0.1
mysql-connector-python==8.2.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 数据处理
opencv-python==4.8.1.78
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

# 自然语言处理
transformers==4.35.2
jieba==0.42.1
nltk==3.8.1

# 向量数据库
asyncpg==0.29.0
sentence-transformers==2.2.2
pgvector==0.2.4

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
